Split Records and Lanes out of ui.rs into their own modules
ui.rs had grown to hold three separate concerns: the builder (Ui,
Draw, ElementMut) that a build actually writes through, and the two
bookkeeping structures (Records, the lane machinery) that a build
writes into but never touches by name. Splitting those out mirrors
store.rs, which already stood alone the same way - lanes.rs and
records.rs now do too, and ui.rs is left holding only the builder
types themselves, roughly half its former size.

records.rs: Records, Watcher, Binding, Elements, and the ChangedFn/
BuildFn contracts a watcher or a binding is built from.

lanes.rs: Lane, Travel, and Lanes - the opaque wrapper Draw and
Records::draw_parts hand out, so Lane itself (pub(crate)) never has
to leave this file to be named anywhere. insert_travel is the one
piece both transition (which already has an Accessor, having read a
base with it) and insert_lane (which builds one from a field closure,
for transition_from/Draw::transition_from) go through, so the Travel
construction itself is written once.

Nothing about the public API moved except where these types live:
`fynix_mock::ui::Records`/`BuildFn`/`ChangedFn` are now
`fynix_mock::records::*`; everything Ui/Draw/ElementMut themselves
expose is unchanged.
