+++ crates/fynix_mock/src/ui.rs
@@ -9,360 +9,19 @@
 //! a bare node, because a node nothing owns could never be patched.
 
 use alloc::boxed::Box;
-use alloc::vec::Vec;
-use core::any::Any;
 use core::marker::PhantomData;
-
-use hashbrown::{HashMap, HashSet};
-use typarena::type_table::TypeTable;
 
 use crate::composer::Composer;
 use crate::element::Element;
 use crate::host::Host;
-use crate::lenz::{Accessor, Cursor, FieldId, FieldPath, Identity};
+use crate::lanes::{Lanes, insert_lane, insert_travel};
+use crate::lenz::{Cursor, FieldPath, Identity};
+use crate::records::{
+    Binding, BuildFn, ChangedFn, Elements, Records, Watcher,
+};
 use crate::store::Store;
 use crate::style::StyledElem;
 use crate::transition::Transition;
-
-/// Predicate over the world, polled once per flush.
-///
-/// `FnMut` because a predicate usually diffs against what it last saw,
-/// and that value has to live somewhere. It must be called exactly
-/// once per flush: a stateful predicate consumes its own signal.
-pub trait ChangedFn<H: Host>:
-    FnMut(&H::World, H::Node) -> bool + Send + Sync + 'static
-{
-}
-
-impl<H: Host, F> ChangedFn<H> for F where
-    F: FnMut(&H::World, H::Node) -> bool + Send + Sync + 'static
-{
-}
-
-/// Rebuilds the subtree under a node.
-pub trait BuildFn<H: Host>:
-    for<'a> Fn(&mut Ui<'a, H>) + Send + Sync + 'static
-{
-}
-
-impl<H: Host, F> BuildFn<H> for F where
-    F: for<'a> Fn(&mut Ui<'a, H>) + Send + Sync + 'static
-{
-}
-
-// The boxed forms the kernel stores.
-type BoxedChanged<H> = Box<
-    dyn FnMut(&<H as Host>::World, <H as Host>::Node) -> bool
-        + Send
-        + Sync,
->;
-type BoxedBuild<H> =
-    Box<dyn for<'a> Fn(&mut Ui<'a, H>) + Send + Sync>;
-
-/// Writes one field of a mounted element, then pushes it out.
-///
-/// It takes the whole table rather than one element, because only the
-/// closure still knows which type to ask for: it was made where that
-/// was still in hand, in [`ElementMut::bind`].
-type BoxedApply<H> = Box<
-    dyn Fn(
-            &mut Elements<H>,
-            &mut <H as Host>::World,
-            <H as Host>::Node,
-            &mut Store<H>,
-        ) + Send
-        + Sync,
->;
-
-/// A field binding: when `changed` fires, the field is written and
-/// patched.
-pub struct Binding<H: Host> {
-    pub(crate) changed: BoxedChanged<H>,
-    pub(crate) apply: BoxedApply<H>,
-}
-
-/// A field on its way somewhere, kept beside the element rather than
-/// in it.
-///
-/// The element keeps the value the cascade gave it, the *base*; a lane
-/// keeps what the backend is showing, and pushes it through the
-/// element's own patch by swapping it in for the length of one call.
-pub(crate) trait Lane<H: Host>: Send + Sync {
-    /// Point this lane somewhere new: a boxed `Option<T>`, `None` to
-    /// release back to the base. Another type is ignored.
-    fn aim(&mut self, target: &mut dyn Any);
-
-    /// Advance by `dt` and push what it reached. `false` once it has
-    /// nothing left to say and the base is what shows.
-    fn advance(
-        &mut self,
-        dt: f32,
-        elements: &mut Elements<H>,
-        world: &mut H::World,
-        node: H::Node,
-        store: &mut Store<H>,
-    ) -> bool;
-}
-
-/// One lane, with the types it was made from still in hand.
-struct Travel<H: Host, E, T: 'static> {
-    accessor: Accessor<E, T>,
-    hops: Vec<FieldId>,
-    transition: Transition<T>,
-    /// What the backend is showing.
-    shown: T,
-    /// Where this leg started.
-    from: T,
-    /// Where it is heading: the target, or the base once released.
-    heading: T,
-    /// The override, or `None` while released.
-    target: Option<T>,
-    elapsed: f32,
-    host: PhantomData<fn() -> H>,
-}
-
-impl<H, E, T> Lane<H> for Travel<H, E, T>
-where
-    H: Host,
-    E: Element<H> + Send + Sync,
-    T: PartialEq + Clone + Send + Sync + 'static,
-{
-    fn aim(&mut self, target: &mut dyn Any) {
-        if let Some(target) = target.downcast_mut::<Option<T>>() {
-            self.target = target.take();
-        }
-    }
-
-    fn advance(
-        &mut self,
-        dt: f32,
-        elements: &mut Elements<H>,
-        world: &mut H::World,
-        node: H::Node,
-        store: &mut Store<H>,
-    ) -> bool {
-        let Some(element) = elements.get_mut::<E>(&node) else {
-            return false;
-        };
-        let Some(base) = (self.accessor.get)(element) else {
-            return false;
-        };
-
-        // The base moves under a running leg whenever a binding writes
-        // it, so where this is heading is worked out afresh each time.
-        let heading =
-            self.target.clone().unwrap_or_else(|| base.clone());
-
-        if heading != self.heading {
-            self.from = self.shown.clone();
-            self.heading = heading;
-            self.elapsed = 0.0;
-        }
-
-        if self.shown == self.heading {
-            // Released and arrived: the base already shows.
-            if self.target.is_none() {
-                return false;
-            }
-        } else {
-            self.elapsed += dt;
-            self.shown = if self.transition.done(self.elapsed) {
-                self.heading.clone()
-            } else {
-                (self.transition.lerp)(
-                    &self.from,
-                    &self.heading,
-                    self.transition.at(self.elapsed),
-                )
-            };
-        }
-
-        // Pushed even when it did not move, so that a binding writing
-        // the base this same flush cannot be the last word.
-        let Some(field) = (self.accessor.get_mut)(element) else {
-            return false;
-        };
-        let base = core::mem::replace(field, self.shown.clone());
-
-        element.patch(world, node, &self.hops, store);
-
-        if let Some(field) = (self.accessor.get_mut)(element) {
-            *field = base;
-        }
-        true
-    }
-}
-
-/// Every field currently travelling rather than snapped to its base,
-/// keyed like the bindings: one per field, so a second lane on the
-/// same one replaces rather than doubles.
-///
-/// Opaque on purpose: [`Lane`] is `pub(crate)`, so this is what a
-/// borrow of the table looks like to anything outside this crate -
-/// [`Draw`] holds one directly, the same way it holds a [`Store`],
-/// without either leaking what a lane actually is.
-pub struct Lanes<H: Host>(
-    HashMap<(H::Node, FieldId), Box<dyn Lane<H>>>,
-);
-
-impl<H: Host> Default for Lanes<H> {
-    fn default() -> Self {
-        Self(HashMap::new())
-    }
-}
-
-impl<H: Host> Lanes<H> {
-    pub(crate) fn insert(
-        &mut self,
-        node: H::Node,
-        key: FieldId,
-        lane: Box<dyn Lane<H>>,
-    ) {
-        self.0.insert((node, key), lane);
-    }
-
-    pub(crate) fn get_mut(
-        &mut self,
-        node: H::Node,
-        key: FieldId,
-    ) -> Option<&mut Box<dyn Lane<H>>> {
-        self.0.get_mut(&(node, key))
-    }
-
-    pub(crate) fn retain(
-        &mut self,
-        mut keep: impl FnMut(H::Node) -> bool,
-    ) {
-        self.0.retain(|(node, _), _| keep(*node));
-    }
-
-    pub(crate) fn iter_mut(
-        &mut self,
-    ) -> impl Iterator<Item = (H::Node, &mut Box<dyn Lane<H>>)> {
-        self.0.iter_mut().map(|((node, _), lane)| (*node, lane))
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
-/// The shared insides of [`ElementMut::transition_from`] and
-/// [`Draw::transition_from`] - only where the lane table comes from
-/// differs between them.
-fn insert_lane<H, E, P>(
-    lanes: &mut Lanes<H>,
-    node: H::Node,
-    field: impl FnOnce(Cursor<Identity<E>>) -> Cursor<P>,
-    base: P::Target,
-    transition: Transition<P::Target>,
-) where
-    H: Host,
-    E: Element<H> + Send + Sync,
-    P: FieldPath<Source = E>,
-    P::Target: PartialEq + Clone + Send + Sync,
-{
-    let cursor = field(Cursor::new());
-    let accessor = cursor.accessor();
-
-    lanes.insert(
-        node,
-        cursor.key(),
-        Box::new(Travel::<H, E, P::Target> {
-            accessor,
-            hops: cursor.hops(),
-            transition,
-            shown: base.clone(),
-            from: base.clone(),
-            heading: base,
-            target: None,
-            elapsed: 0.0,
-            host: PhantomData,
-        }),
-    );
-}
-
-/// A watcher rooted at a node.
-pub struct Watcher<H: Host> {
-    pub(crate) root: H::Node,
-    pub(crate) changed: BoxedChanged<H>,
-    pub(crate) build: BoxedBuild<H>,
-}
-
-/// The elements the kernel built and still owns, by the node each one
-/// took.
-///
-/// A column per element type, so an element comes back as itself: no
-/// erasure to undo, and asking for the wrong type is a miss rather
-/// than a panic.
-pub type Elements<H> = TypeTable<<H as Host>::Node>;
-
-/// What a build registers as it runs, kept beside the world so both
-/// can be borrowed at once.
-pub struct Records<H: Host> {
-    /// Keyed by the whole walk, so two fields of one element are two
-    /// bindings and binding a field twice replaces rather than
-    /// doubles.
-    pub(crate) bindings: HashMap<(H::Node, FieldId), Binding<H>>,
-    /// Keyed like the bindings, and one per field: two overlays on one
-    /// field would each be the last word.
-    pub(crate) lanes: Lanes<H>,
-    pub(crate) elements: Elements<H>,
-    /// Which nodes have a row in `mounts`. The table is keyed by type
-    /// as well as node, so there is no way to ask it what it holds
-    /// without naming a type: this is how a sweep knows what to drop.
-    pub(crate) element_nodes: HashSet<H::Node>,
-    pub(crate) store: Store<H>,
-    /// Watchers declared during a build. They cannot go straight into
-    /// the kernel's list, which is mid-iteration, and must not run
-    /// until the next flush anyway.
-    pub(crate) spawned: Vec<Watcher<H>>,
-}
-
-impl<H: Host> Default for Records<H> {
-    fn default() -> Self {
-        Self {
-            bindings: HashMap::new(),
-            lanes: Lanes::default(),
-            elements: Elements::<H>::new(),
-            element_nodes: HashSet::new(),
-            store: Store::new(),
-            spawned: Vec::new(),
-        }
-    }
-}
-
-impl<H: Host> Records<H> {
-    /// Where every `#[elem(child)]` field's node is recorded. The field
-    /// itself stays private, so a caller that only wants to resolve
-    /// one goes through [`Store::child`] instead of reaching in here.
-    pub fn store(&self) -> &Store<H> {
-        &self.store
-    }
-
-    /// As [`store`](Self::store), for `#[derive(Element)]`'s own
-    /// generated code and anything else driving
-    /// [`Element`](crate::element::Element) by hand - see its
-    /// `build`/`patch`/`despawn`.
-    pub fn store_mut(&mut self) -> &mut Store<H> {
-        &mut self.store
-    }
-
-    /// The store and the lanes together, for
-    /// [`Draw::new`](Draw::new) - what `#[derive(Element)]`'s own
-    /// generated `build` reaches for rather than
-    /// [`store_mut`](Self::store_mut) and a lane accessor called
-    /// separately, which would each want their own exclusive borrow
-    /// of `self` at once.
-    #[doc(hidden)]
-    pub fn draw_parts(&mut self) -> (&mut Lanes<H>, &mut Store<H>) {
-        (&mut self.lanes, &mut self.store)
-    }
-}
 
 /// Builds elements under a parent and records their reactivity.
 pub struct Ui<'a, H: Host> {
@@ -696,20 +355,14 @@ impl<H: Host, E: Element<H>> ElementMut<'_, '_, H, E> {
             return self;
         };
 
-        self.ui.records.lanes.insert(
+        insert_travel(
+            &mut self.ui.records.lanes,
             self.node,
             cursor.key(),
-            Box::new(Travel::<H, E, P::Target> {
-                accessor,
-                hops: cursor.hops(),
-                transition,
-                shown: base.clone(),
-                from: base.clone(),
-                heading: base,
-                target: None,
-                elapsed: 0.0,
-                host: PhantomData,
-            }),
+            accessor,
+            cursor.hops(),
+            base,
+            transition,
         );
         self
     }
