+++ crates/fynix_mock/tests/elements.rs
@@ -7,8 +7,8 @@ use common::{Backend, Label, LabelCursor, World};
 use fynix_mock::element::{Element, ElementVisual, Fields};
 use fynix_mock::host::Host;
 use fynix_mock::lenz::FieldPath;
+use fynix_mock::records::Records;
 use fynix_mock::ui::Draw;
-use fynix_mock::ui::Records;
 
 #[derive(Element)]
 pub struct Icon {
