+++ crates/fynix_mock/tests/siblings.rs
@@ -5,8 +5,8 @@ mod common;
 
 use common::{Label, LabelCursor, World};
 use fynix_mock::element::{Element, ElementVisual};
+use fynix_mock::records::Records;
 use fynix_mock::ui::Draw;
-use fynix_mock::ui::Records;
 
 /// Two labels that already say which is which, so a test can tell one
 /// node from the other without setting anything up.
