+++ crates/fynix_mock/tests/styles.rs
@@ -12,8 +12,8 @@ use common::{Backend, Label, World};
 use fynix_mock::Fynix;
 use fynix_mock::element::Element;
 use fynix_mock::host::Host;
+use fynix_mock::records::Records;
 use fynix_mock::style::{Raw, Style, StyledElem};
-use fynix_mock::ui::Records;
 use fynix_mock::{elem, val};
 
 /// What the cascade produced, for a test that only wants the value.
